refactor nodedelegateimpl to seperate class
